requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
bleach==3.2.1
Django==3.1.4
django-bleach==0.6.1
django-ckeditor==6.0.0
django-js-asset==1.2.2
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
packaging==20.8
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
webencodings==0.5.1
wrapt==1.12.1
